requirements.txt
Django==1.8.6
psycopg2==2.6.1
djangorestframework==3.2.5
djangorestframework-gis==0.9.6
